Refactor imports to remove redundant 'src' prefix and improve code clarity; update deleteComment function to use commentId for better readability.
