+++ src/Backend/routes/user.py
@@ -1,11 +1,11 @@
 import os
 from flask import Blueprint, request, jsonify
-from src.Backend.models import db, User,Companies, Roles, UserRole
+from Backend.models import db, User,Companies, Roles, UserRole
 from sqlalchemy.exc import IntegrityError
 from flask_jwt_extended import jwt_required, create_access_token, get_jwt_identity, get_jwt
 from werkzeug.security import generate_password_hash, check_password_hash
 from werkzeug.utils import secure_filename
-from src.Backend.auth_decorators import role_required
+from Backend.auth_decorators import role_required
 from datetime import datetime, timedelta
 from .password_reset import verification_codes
