+++ src/Frontend/components/Navbar.jsx
@@ -168,7 +168,7 @@ export const Navbar = () => {
                                             boxShadow:
                                                 "0 0 15px rgba(0, 211, 211, 0.8)",
                                         }}
-
+                                        >
                                         {user.avatar ? (
                                             <img
                                                 src={user.avatar}
